requirements.txt
yt-dlp
youtube-transcript-api
streamlit
pandas
chromadb
openai
python-dotenv
tiktoken
